requirements.txt
Django==5.0.7
djangorestframework==3.15.2
psycopg2==2.9.9
python-dotenv==1.0.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
requests~=2.32.3
celery==5.4.0
django-celery-beat==2.6.0
redis==5.0.8
coverage==7.6.0
